Add more 503-SONICWALL sample blocks.
I found many more cases of SonicWALL block pages, evidently from one
user in AS2529 who is stuck behind a network filtering appliance.
However, a variation on the page is still commonly seen (for all
ooniprobes) for the web site www.schwarzreport.org, which seems to use
it on the server side for Tor blocking. The distinction probably will
not cause a problem for us because of the TOR-BLOCKED and NONTOR-BLOCKED
distinction we make.